kaizen-bot/requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0

# Testing
pytest==8.0.0

# Google Calendar integration
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
cryptography==43.0.0
